finalize most envoy logic

+++ src/main/kotlin/sh/miles/pineappleenvoys/util/UnrealizedLocation.kt
@@ -3,9 +3,9 @@ package sh.miles.pineappleenvoys.util
 import org.bukkit.Location
 import org.bukkit.World
 
-data class UnrealizedLocation(val x: Double, val y: Double, val z: Double) {
+data class UnrealizedLocation(val x: Int, val y: Int, val z: Int) {
 
     fun realize(world: World): Location {
-        return Location(world, x, y, z)
+        return Location(world, x.toDouble(), y.toDouble(), z.toDouble())
     }
 }
